Split osu!stable beatmap converter into multiple files
